extricated integration test code
